+++ codec/encode/mp3/source/mp3_enc_utils.c
@@ -3,29 +3,29 @@
  * 32 bit fractional multiplication. Requires 64 bit integer support.
  */
 
-
+#include <stdint.h>
 
 /* Fractional multiply. */
-long mul(long x, long y)
+int32_t mul(int32_t x, int32_t y)
 {
-    return (long)(((long long)x * (long long)y) >> 32);
+    return (int32_t)(((int64_t)x * (int64_t)y) >> 32);
 }
 
 /* Left justified fractional multiply. */
-long muls(long x, long y)
+int32_t muls(int32_t x, int32_t y)
 {
-    return (long)(((long long)x * (long long)y) >> 31);
+    return (int32_t)(((int64_t)x * (int64_t)y) >> 31);
 }
 
 /* Fractional multiply with rounding. */
-long mulr(long x, long y)
+int32_t mulr(int32_t x, int32_t y)
 {
-    return (long)((((long long)x * (long long)y) + 0x80000000) >> 32);
+    return (int32_t)((((int64_t)x * (int64_t)y) + 0x80000000) >> 32);
 }
 
 /* Left justified fractional multiply with rounding. */
-long mulsr(long x, long y)
+int32_t mulsr(int32_t x, int32_t y)
 {
-    return (long)((((long long)x * (long long)y) + 0x40000000) >> 31);
+    return (int32_t)((((int64_t)x * (int64_t)y) + 0x40000000) >> 31);
 }
