+++ codec/encode/speex/libspeex/kiss_fftr.c
@@ -24,7 +24,7 @@ struct kiss_fftr_state{
     kiss_fft_cpx * tmpbuf;
     kiss_fft_cpx * super_twiddles;
 #ifdef USE_SIMD
-    long pad;
+    int32_t pad;
 #endif
 };
